Windows: Fix local file URLs in two autotests
QUrl has been made more stricter when it comes to local files, so
just giving an absolute path with drive letter as URL is not going to
work anymore, as the drive letter will be interpreted as scheme.
Fixed two failing cases related to this issue to use proper URLs.

Task-number: QTBUG-24446
Change-Id: I63afd6b5fd8531bc347316b5dbfc19e4ad7f8115
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>