Add license files mandated by (L)GPL.
Currently only qtbase includes the required LICENSE files, which
are referred to from the code headers. The same files should
be distributed as part of all the modules as well, so that the
(L)GPL licenses are adhered.

These files were wget:d from:

https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LGPL_EXCEPTION.txt
https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LICENSE.LGPL
https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LICENSE.GPL

Not sanitize-commit clean because those files aren't.

Change-Id: I9237ebc08e9a673dea9de81c5ddd348e1a93577e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>