Doc: Replaced \target with \keyword right after \startpage
A \target whose purpose is to link to the top of a
page (and not to a section within a page) works better
as a \keyword, because \target generates a
new html anchor which, in this case, is not tied to
any title element on the page.
A \keyword links to the page itself, as expected.

Task-number: QTBUG-48482
Change-Id: I0d55924f66c26f3fb23a07be1f1d57f2105c879d
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>