Fix QAbstractMessageHandler's docs regarding thread safety
The QMutexLocker's usage was bogus, it didn't do any locking
because it was being created as an anonymous temporary.

It's too late in the game to fix the locking. The user's virtual
function might have it's own locking to workaround this bug and might
re-enter Qtxmlpatterns API causing a dead-lock.

So instead, make the docs consistent with the current behavior.

Change-Id: I4d2bdb1431b44262583abf48f795067ed37281d5
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>