Unbreak ubsan developer-build
GCC's undefined-behavior sanitizer checks that the
declared type of the object is a base class of the
dynamic type of the object on each access to a
member of a class type.

It therefore requires the typeinfo for these types,
which for polymorphic types is emitted in the TU
where the vtable is emitted, too.

QtPatternist::AtomicValue is a polymorphic non-
exported class, so this failed at link-time. Ditto
for the other cases.

Fix by autotest-exporting the classes.

Also, where applicable, de-inline the dtors, so the
vtable (and typeinfo) are pinned to one TU.

Change-Id: I5c47be779a3833433a7108a79a57d61a93ab5494
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>