Significantly improve memory usage in XsdStateMachine::toDFA
We now mark DFA state sets immediately when they are enqueued and
check for that mark before enqueing them. This way, we save a lot
of memory when a set has many states and is encountered often.
In the referenced bug report, there were 60000 copies of a set with
40000 entries, wasting tens of gigabytes of memory.

Task-number: QTBUG-65067
Change-Id: Iec3a04c8badfac29faefbba22c2942ed104effbb
Reviewed-by: Kari Hormi <kari.hormi@qt.io>
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>