Change QSourceLocation::operator!=() to test for inequality
[ChangeLog][QSourceLocation] Corrected QSourceLocation::operator!=()
to test for non-equality, where prior to this it was incorrectly
testing for equality instead.

Also, corrected the QSourceLocation::operator!=() documentation which
was erroneously referencing an unrelated QXmlName class, and updated
the QSourceLocation::operator==() documentation to be consistent with
the rest of the class, as well as broader Qt documentation conventions.

Change-Id: Ia74f84e965f45fd2da3aa42c9c59eba7014b6edb
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>