Catch invalid simple type that was causing a crash
Although the type passes a check that's supposed to find all types
with a simple type, and appears to be of a simple type, its primitive
type is apparently null. This was causing a segmentation fault.

Apparently the code fails to correctly set the primitive type of a
pattern-type instance of a union type. Document the known limitation
by a QEXPECT_FAIL() in the test; I was unable to find where the
primitive type should be set.

Task-number: QTBUG-77620
Change-Id: I56b95ed815366c508f71b8af03b6f2a653d338c5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>