Fix loading of schemas and validation to use resolved URLs
Loading the scheme did compute and record a resolved URL, but didn't
actually pass it on, so a QUrl() of a plain local path (which is
unresolved so doesn't know it's local) isn't recognized by QNAM as
being local when it decides how to handle the request it's been sent.

Likewise, verification's loading of files used the URL it was given,
without resolving, so would fail to read local files it was meant to
validate.

This lead to getting no data from local files.  (That, in turn,
prevented a new test-case I'm about to add from crashing, which the
same change shall fix.)

Change-Id: Icc56dd599897d3498f0838b2f12e3ac46da84519
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>