Fix createTokenLookup.sh and re-run it
This regenerates the tooken lookup code, now without the need for
hand-editing after running the script. The script's copy of a license
header came through shell variable expansion, which broke the $...$
markers; so just cat the relevant lines of the gperf file.

As gperf wants to use size_t for the len parameters of TookenLookup's
hash() and value() methods, pipe its output through a simple sed to
convert size_t to the unsigned int we actually want; and fix a
fall-through marker at the same time.

Change-Id: Ia38c4cd833c6fcaad2371c154e2204c0992aa811
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Albert Astals Cid <albert.astals.cid@kdab.com>