Make a test conditional on its set-up having succeeded
tst_QXmlQuery::setQueryQUrlFailure()'s sub-test "Query via file:/ that
does not have read permissions." depends on the file it tries to read
being unreadable. The test's set-up contained a comment explaining why
we don't QVERIFY() that it succeeded, but went ahead with adding the
test-case regardless. Oddly enough, when the set-up fails to deny
permissions, the test fails because the file can be read, so isn't
found invalid.

Made addition of the test-case conditional on the set-up having
succeeded. This made a long line over-long; and it included a
duplicated long slab of boring code, used in another test, so extract
that piece of code as a lambda and use it twice to save
duplication. The tests were also passing raw C-strings to QUrl(),
which wants a QString, so give it QStringLiteral() in one case and the
existing QString with the relevant name in the other.

Fixes: QTBUG-78560
Change-Id: I2e9ca6fe526edfc021cb4978c81fc9204e047c7f
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>