Update commercial license headers
Updated header.COMM to the files in tqtc-qtxmlpatterns.
Examples, tests or documentation files are not updted.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4173
Change-Id: Ib454a53153affee0d7f3d6b5ff5e7b96f911b3f4
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>